feat: add RepaymentSchedule component
