+++ src/components/BulletLoanRepaymentSchedule.tsx
@@ -1,15 +1,12 @@
 import React, { FunctionComponent, ReactNode, useEffect, useState } from "react";
-import { RepaymentScheduleGraph } from "./RepaymentScheduleGraph";
-import { RepaymentScheduleTable } from "./RepaymentScheduleTable";
-import { Button } from "./Button";
 import {
 	RepaymentScheduleGraphDataPoint,
 	RepaymentScheduleTableDataPoint,
 } from "@credix_types/repaymentschedule.types";
-import { useIntl } from "react-intl";
 import { repaymentSchedule } from "@utils/bullet.utils";
 import { Ratio } from "@credix/credix-client";
 import { generateGraphAndTableData } from "@utils/repayment.utils";
+import { RepaymentSchedule } from "./RepaymentSchedule";
 
 interface BulletLoanRepaymentScheduleProps {
 	principal: number;
@@ -22,8 +19,6 @@ export const BulletLoanRepaymentSchedule: FunctionComponent<BulletLoanRepaymentS
 	principal,
 	financingFee,
 }) => {
-	const intl = useIntl();
-	const [showTable, setShowTable] = useState(false);
 	const [graphData, setGraphData] = useState<RepaymentScheduleGraphDataPoint[]>([]);
 	const [dataSource, setDataSource] = useState<RepaymentScheduleTableDataPoint[]>();
 
@@ -38,30 +33,5 @@ export const BulletLoanRepaymentSchedule: FunctionComponent<BulletLoanRepaymentS
 		}
 	}, [principal, financingFee, timeToMaturity]);
 
-	return (
-		<div>
-			<div className="grid grid-cols-2 gap-x-20 mb-8">
-				<div className="flex flex-col justify-between">
-					<div>
-						<Button onClick={() => setShowTable(!showTable)} type="text">
-							{showTable &&
-								intl.formatMessage({
-									defaultMessage: "Hide table",
-									description: "Deal form: repayment schedule hide table button",
-								})}
-							{!showTable &&
-								intl.formatMessage({
-									defaultMessage: "Show table",
-									description: "Deal form: repayment schedule show table button",
-								})}
-						</Button>
-					</div>
-				</div>
-				<div className="text-right">
-					<RepaymentScheduleGraph data={graphData} />
-				</div>
-			</div>
-			{showTable && <RepaymentScheduleTable dataSource={dataSource} />}
-		</div>
-	);
+	return <RepaymentSchedule graphData={graphData} dataSource={dataSource} />;
 };
