+++ src/consts.ts
@@ -36,69 +36,3 @@ export const dealsRoute: Route = {
 };
 
 export const routes: Route[] = [investWithdrawRoute, dealsRoute];
-
-export const defaultTranches = [
-	{
-		value: "oneTranche",
-		title: "One tranche structure",
-		trancheData: [
-			{
-				name: "Senior",
-				expectedApy: 0.1,
-				value: 1,
-			},
-			{
-				name: "Mezzanine",
-				expectedApy: null,
-				value: null,
-			},
-			{
-				name: "Junior",
-				expectedApy: null,
-				value: null,
-			},
-		],
-	},
-	{
-		title: "Two tranche structure",
-		value: "twoTranche",
-		trancheData: [
-			{
-				name: "Senior",
-				expectedApy: 0.1,
-				value: 0.8,
-			},
-			{
-				name: "Mezzanine",
-				expectedApy: null,
-				value: null,
-			},
-			{
-				name: "Junior",
-				expectedApy: 0.23,
-				value: 0.2,
-			},
-		],
-	},
-	{
-		title: "Three tranche structure",
-		value: "threeTranche",
-		trancheData: [
-			{
-				name: "Senior",
-				expectedApy: 0.1,
-				value: 0.75,
-			},
-			{
-				name: "Mezzanine",
-				expectedApy: 0.17,
-				value: 0.2,
-			},
-			{
-				name: "Junior",
-				expectedApy: 0.34,
-				value: 0.05,
-			},
-		],
-	},
-];
